# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "CoreFeature": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Description",
            "type": "string"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "description"
        ],
        "title": "CoreFeature",
        "type": "object"
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "ImplementationMilestone": {
        "additionalProperties": false,
        "properties": {
          "deliverables": {
            "items": {
              "maxLength": 200,
              "minLength": 1,
              "type": "string"
            },
            "maxItems": 10,
            "minItems": 1,
            "title": "Deliverables",
            "type": "array"
          },
          "title": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "deliverables"
        ],
        "title": "ImplementationMilestone",
        "type": "object"
      },
...
      "PlanRequest": {
        "additionalProperties": false,
        "properties": {
          "idea": {
            "maxLength": 10000,
            "minLength": 1,
            "title": "Idea",
            "type": "string"
          }
        },
        "required": [
          "idea"
        ],
        "title": "PlanRequest",
        "type": "object"
      },
...
      "ProjectPlan": {
        "additionalProperties": false,
        "properties": {
          "core_features": {
            "items": {
              "$ref": "#/components/schemas/CoreFeature"
            },
            "maxItems": 16,
            "minItems": 1,
            "title": "Core Features",
            "type": "array"
          },
          "implementation_milestones": {
            "items": {
              "$ref": "#/components/schemas/ImplementationMilestone"
            },
            "maxItems": 12,
            "minItems": 1,
            "title": "Implementation Milestones",
            "type": "array"
          },
          "product_summary": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Product Summary",
            "type": "string"
          },
          "project_title": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Project Title",
            "type": "string"
          },
          "recommended_stack": {
            "$ref": "#/components/schemas/RecommendedStack"
          },
          "target_users": {
            "items": {
              "maxLength": 200,
              "minLength": 1,
              "type": "string"
            },
            "maxItems": 10,
            "minItems": 1,
            "title": "Target Users",
            "type": "array"
          }
        },
        "required": [
          "project_title",
          "product_summary",
          "target_users",
          "core_features",
          "recommended_stack",
          "implementation_milestones"
        ],
        "title": "ProjectPlan",
        "type": "object"
      },
...
      "RecommendedStack": {
        "additionalProperties": false,
        "properties": {
          "backend": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Backend",
            "type": "string"
          },
          "database": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Database",
            "type": "string"
          },
          "frontend": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Frontend",
            "type": "string"
          },
          "rationale": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Rationale",
            "type": "string"
          }
        },
        "required": [
          "frontend",
          "backend",
          "database",
          "rationale"
        ],
        "title": "RecommendedStack",
        "type": "object"
      },
...
      "UsageView": {
        "additionalProperties": false,
        "properties": {
          "daily_token_budget": {
            "title": "Daily Token Budget",
            "type": "integer"
          },
          "input_token_limit": {
            "title": "Input Token Limit",
            "type": "integer"
          },
          "output_token_limit": {
            "title": "Output Token Limit",
            "type": "integer"
          },
          "requests_per_day": {
            "title": "Requests Per Day",
            "type": "integer"
          },
          "requests_per_minute": {
            "title": "Requests Per Minute",
            "type": "integer"
          },
          "requests_used": {
            "title": "Requests Used",
            "type": "integer"
          },
          "resets_at": {
            "format": "date-time",
            "title": "Resets At",
            "type": "string"
          },
          "tokens_reserved": {
            "title": "Tokens Reserved",
            "type": "integer"
          }
        },
        "required": [
          "requests_used",
          "requests_per_day",
          "requests_per_minute",
          "tokens_reserved",
          "daily_token_budget",
          "input_token_limit",
          "output_token_limit",
          "resets_at"
        ],
        "title": "UsageView",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "AuthenticatedBearer": {
        "description": "Verified provider API token plus server-owned session in OIDC mode; private local credential only in explicit development mode.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "F01 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/plan": {
      "post": {
        "description": "Produces a proposal only. Does not save a project, build code or deploy software.",
        "operationId": "plan_product_v1_plan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectPlan"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Gateway Timeout"
          }
        },
        "security": [
          {
            "AuthenticatedBearer": []
          }
        ],
        "summary": "Create a structured software product plan",
        "tags": [
          "Planning"
        ]
      }
    },
    "/v1/planning/usage": {
      "get": {
        "operationId": "planning_usage_v1_planning_usage_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageView"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "AuthenticatedBearer": []
          }
        ],
        "summary": "Planning Usage",
        "tags": [
          "Planning"
        ]
      }
    },
...
  }
}